pyproject.toml
[project]
name = "blinkdetection"
version = "0.1.0"
description = "Real-time eye blink detection system using transformers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core deep learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",

    # Computer vision
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",
    "pillow>=10.0.0",

    # Data & Configuration
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",

    # Training & Monitoring
    "tensorboard>=2.13.0",
    "tqdm>=4.65.0",

    # Web & API
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0",

    # Utilities
    "requests>=2.31.0",
    "aiofiles>=23.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-pyyaml>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Allow untyped defs for now
ignore_missing_imports = true
